requirements.txt
torch
torchvision
scikit-learn
scipy
matplotlib
seaborn
tqdm
pillow
numpy
flask
flask-cors
werkzeug
requests